Include htm5 doctype in the generated nav file
Epub3 uses xhtml5 for content files. Although the doctype is not mandatory, it’s better to include it. Editors like Sigil will complain that an epub contains files that not well-formed without the doctype.

I’ve therefore add a version check so it’s only inserted for epub v3.0 or greater.